requirements.txt
streamlit
yfinance
pandas
numpy
plotly
yahooquery
newsapi-python
groq